+++ prisma/migrations/20210206155632_/migration.sql
@@ -1,0 +1,9 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `token_version` on the `user` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "user" DROP COLUMN "token_version",
+ADD COLUMN     "tokenVersion" INTEGER NOT NULL DEFAULT 0;