+++ prisma/migrations/20210208233639_fix/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `measurmentQtyId` on the `recipeIngredient` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "recipeIngredient" DROP COLUMN "measurmentQtyId";