+++ prisma/seeders/recipe.seeder.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/ban-ts-ignore */
 import { PrismaClient } from "@prisma/client";
 
 import data from "../seed-data";
@@ -7,6 +8,7 @@ export const recipeSeeder = async (prisma: PrismaClient): Promise<void> => {
 
 	for (const recipe of recipes) {
 		await prisma.recipe.create({
+			// @ts-ignore
 			data: {
 				...recipe,
 			},