+++ src/pages/login.tsx
@@ -3,9 +3,8 @@ import { motion } from "framer-motion";
 import { NextPage } from "next";
 import styled from "@emotion/styled";
 import * as yup from "yup";
-import { useFormik } from "formik";
+import { Field, useFormik, FormikProvider } from "formik";
 import { useRouter } from "next/router";
-import React from "react";
 
 import { Sublink } from "../shared/UIElements";
 import { Mixins } from "../styles/mixins";
@@ -16,7 +15,7 @@ import { setAccessToken } from "../utils/helpers";
 import { withApollo } from "../lib/withApollo";
 import { Wrapper } from "../styles/globals";
 import Button from "../components/Button";
-import { Alert } from "../components/Alert";
+import Alert from "../components/Alert";
 
 interface FormInputs {
 	username: string;
@@ -48,7 +47,7 @@ const LoginScreen: NextPage = () => {
 
 	const router = useRouter();
 
-	const { handleBlur, handleSubmit, handleChange, errors, values, touched } = useFormik({
+	const formik = useFormik({
 		initialValues: {
 			username: "",
 			password: "",
@@ -59,7 +58,7 @@ const LoginScreen: NextPage = () => {
 
 			if (response && response.data) {
 				setAccessToken(response.data.login.accessToken);
-				router.push("/recipes");
+				router.push("/home");
 			}
 		},
 		validationSchema: yup.object().shape({
@@ -85,36 +84,21 @@ const LoginScreen: NextPage = () => {
 			</LoginScreenStyles.TopSection>
 			<LoginScreenStyles.BottomSection>
 				<Wrapper>
-					<Form handleSubmit={handleSubmit}>
-						<FormInput
-							type="text"
-							name="username"
-							onChange={handleChange}
-							touched={touched.username}
-							value={values.username}
-							error={errors.username}
-							onBlur={handleBlur}
-						/>
+					<FormikProvider value={formik}>
+						<Form handleSubmit={formik.handleSubmit}>
+							<Field name="username" component={FormInput} />
+							<Mixins.Spacer height="20px" />
+							<Field type="password" name="password" component={FormInput} />
+							<Mixins.Flex flexEnd style={{ paddingTop: 10, paddingBottom: 24 }}>
+								<Sublink href="/login" linkedText="Forgot Password?" />
+							</Mixins.Flex>
+							<Button type="submit" label="Login" size="full" primary />
 
-						<FormInput
-							type="password"
-							name="password"
-							onChange={handleChange}
-							touched={touched.password}
-							value={values.password}
-							error={errors.password}
-							onBlur={handleBlur}
-						/>
-						<Mixins.Flex flexEnd style={{ paddingTop: 10, paddingBottom: 24 }}>
-							<Sublink href="/login" linkedText="Forgot Password?" />
-						</Mixins.Flex>
-						<Button type="submit" label="Login" size="full" primary />
-
-						<Mixins.Flex center style={{ paddingTop: 26 }}>
-							<Sublink href="/register" unlinkedText="Don't have an account? " linkedText="Signup." />
-						</Mixins.Flex>
-					</Form>
-
+							<Mixins.Flex center style={{ paddingTop: 26 }}>
+								<Sublink href="/register" unlinkedText="Don't have an account? " linkedText="Signup." />
+							</Mixins.Flex>
+						</Form>
+					</FormikProvider>
 					<Mixins.Flex center style={{ paddingTop: 26 }}>
 						{error && <Alert message={errorMessage} severity="error" />}
 					</Mixins.Flex>