+++ src/utils/variables.ts
@@ -3,10 +3,12 @@ const colors = {
 	lightSecondary: "#F8F5F2",
 	lightTertiary: "#CE4B2E",
 	lightQuaternary: "#222632",
+	lightQuinary: "#F2ECE6",
 	darkPrimary: "",
 	darkSecondary: "",
 	darkTertiary: "",
 	darkQuaternary: "",
+	darkQuinary: "",
 	accent: "",
 };