+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.7](https://github.com/darkristy/recipe-app/compare/v0.1.6...v0.1.7) (2021-02-12)
+
+
+### Bug Fixes
+
+* fixed mistake in seeders ([da985cb](https://github.com/darkristy/recipe-app/commit/da985cba42ee5d35a8ba9000be7e9b287718e5d8))
+
+### [0.1.6](https://github.com/darkristy/recipe-app/compare/v0.1.5...v0.1.6) (2021-02-12)
+
+
+### Features
+
+* add mini button component ([87bcf83](https://github.com/darkristy/recipe-app/commit/87bcf8311b9e7ef64ec280c00f40bc7a8563c243))
+* add new mixins ([5c0fc26](https://github.com/darkristy/recipe-app/commit/5c0fc266ae0d7b36f552d7f92ee76d9bc96e7bec))
+* add new theme colors ([c9173ef](https://github.com/darkristy/recipe-app/commit/c9173ef97bc3ddac9d02e4a650111e87e640298c))
+* new recipe page built just need server logic ([21f2ae4](https://github.com/darkristy/recipe-app/commit/21f2ae434b53b1ebe74eb0b268d50d0ccf72e4e8))
+
+### [0.1.5](https://github.com/darkristy/recipe-app/compare/v0.1.4...v0.1.5) (2021-02-07)
+
+
+### Features
+
+* add main pages ([96e0b33](https://github.com/darkristy/recipe-app/commit/96e0b33fe6dd447ecb217c9e8c64aa3b57d53e95))
+
 ### [0.1.4](https://github.com/darkristy/recipe-app/compare/v0.1.3...v0.1.4) (2021-02-06)